feat(db): nullable owner_id, full-field create/update endpoints
projects/processes owner_id now nullable with FK to users; create/update
requests expose all record fields (owner_id, project_id, process_id,
task_id, resource_id). Regenerate sqlc models and swagger docs.

+++ internal/project_mgmt/project/repository/query/projects.sql
@@ -15,7 +15,7 @@ VALUES (
   @start_date::date,
   @end_date::date,
   @priority::bigint,
-  @owner_id::bigint
+  @owner_id
 )
 RETURNING *;
 
@@ -53,7 +53,7 @@ SET
 	priority = @priority::bigint,
 	start_date = @start_date,
 	end_date = @end_date,
-  owner_id = @owner_id::bigint,
+  owner_id = @owner_id,
 	updated_at = NOW()
 WHERE deleted_at IS NULL
 	AND id = @project_id::bigint
